Adjusted DeletePostHandler auth for post functions
Added a check for if the user is logged in

+++ main.py
@@ -413,14 +413,17 @@ class DeletePostHandler(BlogHandler):
     def post(self, id):
         article = Article.get_by_id(int(id))
         isLoggedIn = self.isLoggedIn()
-        if (isLoggedIn and article.user == isLoggedIn):
-            article.delete()
-            self.render('message.html',
-                        message="Post deletion successful.",
-                        isLoggedIn=isLoggedIn)
+        if isLoggedIn:
+            if (isLoggedIn and article.user == isLoggedIn):
+                article.delete()
+                self.render('message.html',
+                            message="Post deletion successful.",
+                            isLoggedIn=isLoggedIn)
+            else:
+                self.render("message.html",
+                            error="That's not permitted")
         else:
-            self.render("message.html",
-                        error="That's not permitted")
+            self.redirect("/login")
 
 
 class LikePostHandler(BlogHandler):
